pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mallcop"
version = "0.1.0"
description = "Security monitoring for small cloud operators"
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
dependencies = [
    "click>=8.0",
    "defusedxml>=0.7",
    "pyyaml>=6.0",
    "requests>=2.28",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
]

[project.scripts]
mallcop = "mallcop.cli:cli"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-m 'not live'"
markers = [
    "live: requires real API credentials (ANTHROPIC_API_KEY, AZURE_*)",
]
